[refactor] Create a file for the BasicChecker in pylint.checker.base
